renamed a few dir's, and removed not needed code
