--- pwa/src/modules/network.js
@@ -1,9 +1,0 @@
-// make an abstract class for network restful network interactions.
-/*
-
-    .get()
-    .post()
-    .delete()
-    
-
- */
