--- pwa/src/store/user/index.js
@@ -1,6 +1,0 @@
-import user from './reducer';
-
-export * from './actions';
-export * from './types';
-
-export default user;
