--- pwa/src/store/user/types.js
@@ -1,4 +1,0 @@
-/*** ACTION TYPES ***/
-export const USER = 'USER';
-export const SOCKET = 'SOCKET';
-export const CLIENT = 'CLIENT';
