+++ pwa/src/utils/registerServiceWorker.js
