+++ pwa/src/views/about/index.js
