+++ pwa/src/views/signup/index.js
@@ -1,15 +1,12 @@
-import { push } from 'react-router-redux'
-import { bindActionCreators } from 'redux'
-import { connect } from 'react-redux'
+import { push } from 'react-router-redux';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
 
 import SignUpView from './view';
-
-import { createClient } from '../../store/user';
 
 const mapStateToProps = state => ({})
 
 const mapDispatchToProps = dispatch => bindActionCreators({
-    createClient,
     toDashboard: () => push('/dashboard'),
     toHome:      () => push('/')
 }, dispatch)
